updating constants and refactoring functions
